requirements.txt
python-gitlab>=5.3,<6.0
python-dotenv>=1.0,<2.0
cryptography>=44.0,<45.0
pydantic>=2.0.0,<3.0.0
pandas>=2.0.0,<3.0.0
pytest>=7.0.0
setuptools>=75.8,<76.0